refactor natspec script
